Improve globbing behavior for single-file compilation
I recently migrated from Aniseed to nfnl and everything went great
(thanks so much for these tools, by the way!). Then I started migrating
my project-local config files for work, at which point I discovered that
nfnl seemed to be hanging whenever I tried to save an `.nfnl.fnl` file.

The behavior of globbing all compilable files on save works fine for a
small to medium-size repo, but it's not great for the `.nfnl.fnl` case
in larger codebases with `node_modules` directories, Cargo caches, and
so on. We can work around this in the single-file case by using
`glob2regpat` and testing each glob against the file's path directly.

This isn't a perfect solution since it doesn't help with macros, but
that seems like a bigger design challenge that's luckily not relevant
for my `.nfnl.fnl` files, and this at least seems like it shouldn't
hurt. It also lets us get rid of the `file-exists-on-disk?` flag.